+++ src/main/kotlin/com/github/trcdevelopers/clayium/common/blocks/clay/ItemBlockEnergizedClay.kt
@@ -1,8 +1,8 @@
 package com.github.trcdevelopers.clayium.common.blocks.clay
 
-import com.github.trcdevelopers.clayium.common.ClayEnergy
+import com.github.trcdevelopers.clayium.common.clayenergy.ClayEnergy
 import com.github.trcdevelopers.clayium.common.Clayium
-import com.github.trcdevelopers.clayium.common.interfaces.IEnergizedClay
+import com.github.trcdevelopers.clayium.common.clayenergy.IEnergizedClay
 import com.github.trcdevelopers.clayium.common.util.CUtils
 import net.minecraft.client.resources.I18n
 import net.minecraft.client.util.ITooltipFlag
