+++ src/main/kotlin/com/github/trcdevelopers/clayium/common/recipe/IMachineRecipe.kt
@@ -1,6 +1,6 @@
 package com.github.trcdevelopers.clayium.common.recipe
 
-import com.github.trcdevelopers.clayium.common.ClayEnergy
+import com.github.trcdevelopers.clayium.common.clayenergy.ClayEnergy
 
 interface IMachineRecipe {
     val requiredTicks: Int
